Rename maxsum into asynchronous maxsum
